+++ tests/testthat/test-validate.R
@@ -63,8 +63,10 @@ test_that("pacs::pac_health online", {
   skip_if_offline()
   expect_true(is.logical(pac_health("dplyr")))
   expect_true(is.logical(pac_health("dplyr", scope = c("ERROR"))))
-  expect_true(isFALSE(pac_health("dplyr", version = "0.8.0")))
-  expect_true(isFALSE(pac_health("dplyr", at = as.Date("2019-02-14"))))
+  expect_true(isFALSE(pac_health("dplyr", version = "0.8.0")) ||
+                isNA(pac_health("dplyr", version = "0.8.0")))
+  expect_true(isFALSE(pac_health("dplyr", at = as.Date("2019-02-14"))) ||
+                isNA(pac_health("dplyr", at = as.Date("2019-02-14"))))
 })
 
 test_that("pacs::pac_checkred", {